LeNet/requirements.txt
torch
torchvision
pytorch_lightning
pillow
